+++ ChangeLog.md
@@ -1,5 +1,8 @@
 2021-08
 
+- enable spectrogram for any zoom.
+  However, for now, destroying the spectrogram when zoom is changed
+  as it seems the zoom change doesn't propagate to the spectrogram
 - allow some spectrogram params
 - recreate wavesurfer upon loading file.
   This seems to fix some misbehavior when trying to reuse the